Codeclimate вот он я
